start and end time should be float

+++ app/concepts/api/v1/videos/contract/create.rb
@@ -11,8 +11,8 @@ module Api
         validation do
           params do
             required(:file).filled
-            required(:start).filled(:integer)
-            required(:end).filled(:integer)
+            required(:start).filled(:float)
+            required(:end).filled(:float)
           end
 
           rule(:start) do
